add e2e tcp_check
