+++ tcp_check/tests/conftest.py
@@ -1,0 +1,18 @@
+# (C) Datadog, Inc. 2019
+# All rights reserved
+# Licensed under Simplified BSD License (see LICENSE)
+import pytest
+
+from datadog_checks.tcp_check import TCPCheck
+
+from . import common
+
+
+@pytest.fixture(scope='session')
+def dd_environment():
+    yield common.INSTANCE
+
+
+@pytest.fixture
+def check():
+    return TCPCheck(common.CHECK_NAME, {}, {})
